klar med uppgift 1 lek 3

+++ Lek3 Uppgift1/src/Bil.java
@@ -12,7 +12,7 @@ public class Bil extends Vagfordon {
         }
     }
 
-    public double getAntalSittplatser() {
+    public int getAntalSittplatser() {
         return sittplatser;
     }
